Refactor: Move MicrophoneProvider to services package
