+++ composeApp/src/androidMain/kotlin/org/noiseplanet/noisecapture/AndroidPlatformModule.kt
@@ -10,13 +10,13 @@ import org.koin.core.module.Module
 import org.koin.dsl.module
 import org.noiseplanet.noisecapture.audio.AndroidAudioSource
 import org.noiseplanet.noisecapture.audio.AudioSource
-import org.noiseplanet.noisecapture.audio.mic.AndroidMicrophoneProvider
-import org.noiseplanet.noisecapture.audio.mic.MicrophoneProvider
 import org.noiseplanet.noisecapture.audio.player.AndroidAudioPlayer
 import org.noiseplanet.noisecapture.audio.player.AudioPlayer
 import org.noiseplanet.noisecapture.log.Logger
 import org.noiseplanet.noisecapture.services.audio.AndroidAudioRecordingService
+import org.noiseplanet.noisecapture.services.audio.AndroidMicrophoneProviderService
 import org.noiseplanet.noisecapture.services.audio.AudioRecordingService
+import org.noiseplanet.noisecapture.services.audio.MicrophoneProviderService
 import org.noiseplanet.noisecapture.services.location.AndroidUserLocationProvider
 import org.noiseplanet.noisecapture.services.location.UserLocationProvider
 import org.noiseplanet.noisecapture.services.measurement.AndroidRecordingService
@@ -71,8 +71,8 @@ val platformModule: Module = module {
         AndroidFilePickerEventBus()
     }
 
-    single<MicrophoneProvider> {
-        AndroidMicrophoneProvider()
+    single<MicrophoneProviderService> {
+        AndroidMicrophoneProviderService()
     }
 
     factory<AudioPlayer> { (filePath: String) ->
