+++ composeApp/src/androidMain/kotlin/org/noiseplanet/noisecapture/services/audio/AndroidAudioRecordingService.kt
@@ -6,7 +6,6 @@ import android.media.MediaRecorder
 import android.os.Build
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import org.noiseplanet.noisecapture.audio.mic.MicrophoneProvider
 import org.noiseplanet.noisecapture.log.Logger
 import org.noiseplanet.noisecapture.services.storage.FileSystemService
 import org.noiseplanet.noisecapture.util.getInputDevice
@@ -22,7 +21,7 @@ class AndroidAudioRecordingService : AudioRecordingService, KoinComponent {
     private val logger: Logger by injectLogger()
     private val context: Context by inject()
     private val fileSystemService: FileSystemService by inject()
-    private val microphoneProvider: MicrophoneProvider by inject()
+    private val microphoneProvider: MicrophoneProviderService by inject()
 
     private var mediaRecorder: MediaRecorder? = null
     private var outputFile: File? = null
