+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/model/dao/MicrophoneCalibrationProfile.kt
@@ -1,7 +1,6 @@
 package org.noiseplanet.noisecapture.model.dao
 
 import kotlinx.serialization.Serializable
-import org.noiseplanet.noisecapture.audio.mic.MicrophoneType
 
 
 /**
