+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/home/HomeMicrophoneSetupViewModel.kt
@@ -10,8 +10,8 @@ import noisecapture.composeapp.generated.resources.home_mic_setup_calibrate_butt
 import noisecapture.composeapp.generated.resources.home_mic_setup_recalibrate_button
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import org.noiseplanet.noisecapture.audio.mic.MicrophoneProvider
 import org.noiseplanet.noisecapture.model.dao.MicrophoneCalibrationProfile
+import org.noiseplanet.noisecapture.services.audio.MicrophoneProviderService
 import org.noiseplanet.noisecapture.ui.components.button.NCButtonColors
 import org.noiseplanet.noisecapture.ui.components.button.NCButtonStyle
 import org.noiseplanet.noisecapture.ui.components.button.NCButtonViewModel
@@ -33,7 +33,7 @@ class HomeMicrophoneSetupViewModel : ViewModel(), KoinComponent {
 
     // - Properties
 
-    private val microphoneProvider: MicrophoneProvider by inject()
+    private val microphoneProvider: MicrophoneProviderService by inject()
 
     val viewState: StateFlow<ViewState> = microphoneProvider.currentCalibrationProfile
         .map { calibrationProfile ->
