+++ composeApp/src/iosMain/kotlin/org/noiseplanet/noisecapture/IOSPlatformModule.kt
@@ -9,13 +9,13 @@ import org.koin.core.module.Module
 import org.koin.dsl.module
 import org.noiseplanet.noisecapture.audio.AudioSource
 import org.noiseplanet.noisecapture.audio.IOSAudioSource
-import org.noiseplanet.noisecapture.audio.mic.IOSMicrophoneProvider
-import org.noiseplanet.noisecapture.audio.mic.MicrophoneProvider
 import org.noiseplanet.noisecapture.audio.player.AudioPlayer
 import org.noiseplanet.noisecapture.audio.player.IOSAudioPlayer
 import org.noiseplanet.noisecapture.log.Logger
 import org.noiseplanet.noisecapture.services.audio.AudioRecordingService
 import org.noiseplanet.noisecapture.services.audio.IOSAudioRecordingService
+import org.noiseplanet.noisecapture.services.audio.IOSMicrophoneProviderService
+import org.noiseplanet.noisecapture.services.audio.MicrophoneProviderService
 import org.noiseplanet.noisecapture.services.location.IOSUserLocationProvider
 import org.noiseplanet.noisecapture.services.location.UserLocationProvider
 import org.noiseplanet.noisecapture.services.storage.FileSystemService
@@ -68,7 +68,7 @@ val platformModule: Module = module {
         IOSAudioPlayer(filePath)
     }
 
-    single<MicrophoneProvider> {
-        IOSMicrophoneProvider()
+    single<MicrophoneProviderService> {
+        IOSMicrophoneProviderService()
     }
 }
