+++ composeApp/src/wasmJsMain/kotlin/org/noiseplanet/noisecapture/JSPlatformModule.kt
@@ -8,13 +8,13 @@ import org.koin.core.module.Module
 import org.koin.dsl.module
 import org.noiseplanet.noisecapture.audio.AudioSource
 import org.noiseplanet.noisecapture.audio.JsAudioSource
-import org.noiseplanet.noisecapture.audio.mic.JSMicrophoneProvider
-import org.noiseplanet.noisecapture.audio.mic.MicrophoneProvider
 import org.noiseplanet.noisecapture.audio.player.AudioPlayer
 import org.noiseplanet.noisecapture.audio.player.JSAudioPlayer
 import org.noiseplanet.noisecapture.log.Logger
 import org.noiseplanet.noisecapture.services.audio.AudioRecordingService
 import org.noiseplanet.noisecapture.services.audio.JSAudioRecordingService
+import org.noiseplanet.noisecapture.services.audio.JSMicrophoneProviderService
+import org.noiseplanet.noisecapture.services.audio.MicrophoneProviderService
 import org.noiseplanet.noisecapture.services.location.UserLocationProvider
 import org.noiseplanet.noisecapture.services.location.WasmJSUserLocationProvider
 import org.noiseplanet.noisecapture.services.storage.FileSystemService
@@ -55,7 +55,7 @@ val platformModule: Module = module {
         JSAudioPlayer(filePath)
     }
 
-    single<MicrophoneProvider> {
-        JSMicrophoneProvider()
+    single<MicrophoneProviderService> {
+        JSMicrophoneProviderService()
     }
 }
